Struct smallbox::SmallBox

pub struct SmallBox<T: ?Sized, Space> { /* fields omitted */ }

An optimized box that store value on stack or heap depending on its size

Methods

impl<T: ?Sized, Space> SmallBox<T, Space>

pub fn new(val: T) -> SmallBox<T, Space> where
T: Sized,

Box value on stack or heap depending on its size.

Example

use smallbox::SmallBox;
use smallbox::space::*;

let small: SmallBox<_, S4> = SmallBox::new([0usize; 2]);
let large: SmallBox<_, S4> = SmallBox::new([1usize; 8]);

assert_eq!(small.len(), 2);
assert_eq!(large[7], 1);

assert!(large.is_heap() == true);

pub fn resize<ToSpace>(self) -> SmallBox<T, ToSpace>

Change the capacity of SmallBox

This method may move stack-allocated data to heap if the inline space is not sufficient. Once the data is stored on heap, it'll never be moved again.

Example

use smallbox::SmallBox;
use smallbox::space::{S2, S4};

let s: SmallBox::<_, S4> = SmallBox::new([0usize; 4]);
let m: SmallBox::<_, S2> = s.resize();

pub fn is_heap(&self) -> bool

Returns true if the data is heap-allocated