Struct smallbox::SmallBox

pub struct SmallBox<T: ?Sized, Space> { /* private fields */ }

An optimized box that store value on stack or on heap depending on its size

Implementations

impl<T: ?Sized, Space> SmallBox<T, Space>

pub fn new(val: T) -> SmallBox<T, Space>where T: Sized,

Box value on stack or on heap depending on its size.

Example

use smallbox::space::*;
use smallbox::SmallBox;

let small: SmallBox<_, S4> = SmallBox::new([0usize; 2]);
let large: SmallBox<_, S4> = SmallBox::new([1usize; 8]);

assert_eq!(small.len(), 2);
assert_eq!(large[7], 1);

assert!(large.is_heap() == true);

pub fn resize<ToSpace>(self) -> SmallBox<T, ToSpace>

Change the capacity of SmallBox.

This method may move stack-allocated data from stack to heap when inline space is not sufficient. And once the data is moved to heap, it’ll never be moved again.

Example

use smallbox::space::S2;
use smallbox::space::S4;
use smallbox::SmallBox;

let s: SmallBox<_, S4> = SmallBox::new([0usize; 4]);
let m: SmallBox<_, S2> = s.resize();

pub fn is_heap(&self) -> bool

Returns true if data is allocated on heap.

Example

use smallbox::space::S1;
use smallbox::SmallBox;

let stacked: SmallBox<usize, S1> = SmallBox::new(0usize);
assert!(!stacked.is_heap());

let heaped: SmallBox<(usize, usize), S1> = SmallBox::new((0usize, 1usize));
assert!(heaped.is_heap());

pub fn into_inner(self) -> Twhere T: Sized,

Consumes the SmallBox and returns ownership of the boxed value

Examples

use smallbox::space::S1;
use smallbox::SmallBox;

let stacked: SmallBox<_, S1> = SmallBox::new([21usize]);
let val = stacked.into_inner();
assert_eq!(val[0], 21);

let boxed: SmallBox<_, S1> = SmallBox::new(vec![21, 56, 420]);
let val = boxed.into_inner();
assert_eq!(val[1], 56);

impl<Space> SmallBox<dyn Any, Space>

pub fn downcast<T: Any>(self) -> Result<SmallBox<T, Space>, Self>

Attempt to downcast the box to a concrete type.

Examples

#[macro_use]
extern crate smallbox;

use core::any::Any;

use smallbox::space::*;
use smallbox::SmallBox;

fn print_if_string(value: SmallBox<dyn Any, S1>) {
    if let Ok(string) = value.downcast::<String>() {
        println!("String ({}): {}", string.len(), string);
    }
}

fn main() {
    let my_string = "Hello World".to_string();
    print_if_string(smallbox!(my_string));
    print_if_string(smallbox!(0i8));
}

impl<Space> SmallBox<dyn Any + Send, Space>

pub fn downcast<T: Any>(self) -> Result<SmallBox<T, Space>, Self>

Attempt to downcast the box to a concrete type.

Examples

#[macro_use]
extern crate smallbox;

use core::any::Any;

use smallbox::space::*;
use smallbox::SmallBox;

fn print_if_string(value: SmallBox<dyn Any, S1>) {
    if let Ok(string) = value.downcast::<String>() {
        println!("String ({}): {}", string.len(), string);
    }
}

fn main() {
    let my_string = "Hello World".to_string();
    print_if_string(smallbox!(my_string));
    print_if_string(smallbox!(0i8));
}